+++ allure-specflow/Allure.SpecFlowPlugin.Tests/app.config
@@ -3,16 +3,16 @@
   <configSections>
     <section name="specFlow" type="TechTalk.SpecFlow.Configuration.ConfigurationSectionHandler, TechTalk.SpecFlow" />
   </configSections>
+
   <specFlow>
     <plugins>
       <add name="Allure" type="Runtime" />
-      <!--<add name="SpecRun" />-->
+      <add name="SpecRun" />
     </plugins>
     <stepAssemblies>
       <stepAssembly assembly="Allure.SpecFlowPlugin" />
     </stepAssemblies>
-    <unitTestProvider name="nunit" />
-    <!--<unitTestProvider name="SpecRun" />-->
+    <unitTestProvider name="SpecRun" />
   </specFlow>
 
   <runtime>
