+++ allure-specflow/Allure.SpecFlowPlugin.Tests/packages.config
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Allure.Commons" version="0.2.7-dev" targetFramework="net462" />
+  <package id="Allure.Commons" version="0.2.23-dev" targetFramework="net462" />
+  <package id="Castle.Core" version="4.1.1" targetFramework="net462" />
   <package id="Microsoft.Extensions.Configuration" version="1.1.2" targetFramework="net462" />
   <package id="Microsoft.Extensions.Configuration.Abstractions" version="1.1.2" targetFramework="net462" />
   <package id="Microsoft.Extensions.Configuration.FileExtensions" version="1.1.2" targetFramework="net462" />
@@ -14,8 +15,11 @@
   <package id="MimeTypesMap" version="1.0.1" targetFramework="net462" />
   <package id="NETStandard.Library" version="1.6.1" targetFramework="net462" />
   <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net462" />
+  <package id="NLog" version="5.0.0-beta09" targetFramework="net462" />
   <package id="NUnit" version="3.7.1" targetFramework="net462" />
-  <package id="SpecFlow" version="2.2.0" targetFramework="net462" />
+  <package id="SpecFlow" version="2.1.0" targetFramework="net462" />
+  <package id="SpecRun.Runner" version="1.5.2" targetFramework="net462" />
+  <package id="SpecRun.SpecFlow" version="1.5.2" targetFramework="net462" />
   <package id="System.AppContext" version="4.3.0" targetFramework="net462" />
   <package id="System.Collections" version="4.3.0" targetFramework="net462" />
   <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="net462" />
