+++ allure-specflow/Allure.SpecFlowPlugin/App.config
@@ -3,7 +3,17 @@
   <configSections>
     <section name="specFlow" type="TechTalk.SpecFlow.Configuration.ConfigurationSectionHandler, TechTalk.SpecFlow" />
   </configSections>
-  
+
+  <specFlow>
+    <!-- For additional details on SpecFlow configuration options see http://go.specflow.org/doc-config -->
+    <!-- For additional details on SpecFlow configuration options see http://go.specflow.org/doc-config -->
+    <!-- use unit test provider SpecRun+NUnit or SpecRun+MsTest for being able to execute the tests with SpecRun and another provider -->
+    <unitTestProvider name="SpecRun" />
+    <plugins>
+      <add name="SpecRun" />
+    </plugins>
+  </specFlow>
+
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
       <dependentAssembly>
